# api-specs/openapi/core-api.yaml
openapi: 3.0.3
info:
  title: オンライン契約システム Core API
  description: |
    スマホ・携帯プランオンライン契約システムのCore API仕様
    
    SCRUM-77: API仕様統一設計 - フロント・バックエンド間インターフェース詳細設計
    Parent Epic: SCRUM-30 - オンライン契約システム
    
    ## 機能概要
    - プラン選択・比較 (3種類)
    - 機種選択・検索 (20機種対応)
    - オプション選択
    - 契約手続き (個人情報、eKYC、電子署名)
    - 決済・配送
    - MNP対応
    
    ## API設計原則
    - RESTful設計
    - JSON形式での通信
    - 統一的なエラーハンドリング
    - セキュアな認証・認可
    - レスポンシブ対応
  version: "1.0.0"
  contact:
    name: オンライン契約システム開発チーム
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.online-contract.example.com/v1
    description: 本番環境
  - url: https://api-staging.online-contract.example.com/v1
    description: ステージング環境
  - url: http://localhost:3000/v1
    description: 開発環境

# セキュリティ設定
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# 共通パス定義
paths:
...
  /contracts:
    post:
      tags:
        - Contracts
      summary: 契約作成
      description: 新規契約を作成（途中保存含む）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractCreateRequest'
      responses:
        '201':
          description: 契約作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    get:
      tags:
        - Contracts
      summary: 契約一覧取得
      description: ユーザーの契約一覧を取得
      parameters:
        - name: status
          in: query
          description: 契約ステータスフィルタ
          required: false
          schema:
            type: string
            enum: [draft, pending, active, cancelled]
      responses:
        '200':
          description: 契約一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  contracts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Contract'
...
  /contracts/templates:
    get:
      tags:
        - Contract Templates
      summary: 契約テンプレート一覧取得
      description: 利用可能な契約テンプレートの一覧を取得
      responses:
        '200':
          description: テンプレート一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContractTemplate'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Contract Templates
      summary: 契約テンプレート作成
      description: 新しい契約テンプレートを作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractTemplateCreateRequest'
      responses:
        '201':
          description: テンプレート作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractTemplate'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /contracts/templates/{templateId}:
    get:
      tags:
        - Contract Templates
      summary: 契約テンプレート詳細取得
      description: 指定テンプレートの詳細情報を取得
      parameters:
        - name: templateId
          in: path
          required: true
          description: テンプレートID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: テンプレート詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractTemplate'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Contract Templates
      summary: 契約テンプレート更新
      description: 契約テンプレートを更新（新バージョン作成）
      parameters:
        - name: templateId
          in: path
          required: true
          description: テンプレートID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractTemplateUpdateRequest'
      responses:
        '200':
          description: テンプレート更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractTemplate'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /contracts/{contractId}/generate:
    post:
      tags:
        - Contract Generation
      summary: 契約書生成
      description: 顧客情報に基づいて動的に契約書を生成
      parameters:
        - name: contractId
          in: path
          required: true
          description: 契約ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractGenerationRequest'
      responses:
        '200':
          description: 契約書生成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractGenerationResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /contracts/{contractId}/sign:
    post:
      tags:
        - Electronic Signature
      summary: 電子署名実行
      description: 契約書に電子署名を実行し、タイムスタンプを付与
      parameters:
        - name: contractId
          in: path
          required: true
          description: 契約ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElectronicSignatureRequest'
      responses:
        '200':
          description: 電子署名成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElectronicSignatureResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /contracts/{contractId}/document:
    get:
      tags:
        - Contract Documents
      summary: 契約書文書取得
      description: 署名済み契約書文書を取得
      parameters:
        - name: contractId
          in: path
          required: true
          description: 契約ID
          schema:
            type: string
            format: uuid
        - name: format
          in: query
          description: 文書形式
          required: false
          schema:
            type: string
            enum: [pdf, html]
            default: pdf
      responses:
        '200':
          description: 契約書文書
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                $ref: '#/components/schemas/ContractDocumentResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /contracts/{contractId}/audit-trail:
    get:
      tags:
        - Contract Audit
      summary: 監査証跡取得
      description: 契約に関する監査証跡を取得
      parameters:
        - name: contractId
          in: path
          required: true
          description: 契約ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 監査証跡
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractAuditTrailResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /contracts/search:
    get:
      tags:
        - Contract Search
      summary: 契約検索
      description: 条件を指定して契約を検索
      parameters:
        - name: customer_name
          in: query
          description: 顧客名
          required: false
          schema:
            type: string
        - name: contract_number
          in: query
          description: 契約番号
          required: false
          schema:
            type: string
        - name: status
          in: query
          description: 契約ステータス
          required: false
          schema:
            type: string
            enum: [draft, pending, signed, active, cancelled]
        - name: date_from
          in: query
          description: 作成日時（開始）
          required: false
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          description: 作成日時（終了）
          required: false
          schema:
            type: string
            format: date
        - name: page
          in: query
          description: ページ番号
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: 1ページあたりの件数
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 検索結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractSearchResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'

  # ========================================
  # eKYC・本人確認
  # ========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT認証トークン
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API キー認証

  # ========================================
  # スキーマ定義
  # ========================================
  schemas:
...
    Plan:
      type: object
      properties:
        id:
          type: string
          example: "plan_basic_001"
        name:
          type: string
          example: "ベーシックプラン"
        description:
          type: string
          example: "月額2,980円の基本プラン"
        monthly_fee:
          type: number
          format: decimal
          example: 2980
        data_capacity:
          type: string
          example: "20GB"
        voice_calls:
          type: string
          example: "無制限"
        sms:
          type: string
          example: "無制限"
        features:
          type: array
          items:
            type: string
        is_active:
          type: boolean
          example: true
        version:
          type: string
          example: "1.0.0"
          description: "プランバージョン"
        parent_plan_id:
          type: string
          example: "plan_basic_001_v1"
          description: "親プランID（更新時の元プラン）"
        is_current_version:
          type: boolean
          example: true
          description: "現在有効なバージョンかどうか"
        effective_start_date:
          type: string
          format: date-time
          example: "2024-12-20T00:00:00Z"
          description: "有効開始日時"
        effective_end_date:
          type: string
          format: date-time
          example: "2025-12-19T23:59:59Z"
          description: "有効終了日時"
        campaign_start_date:
          type: string
          format: date-time
          example: "2024-12-20T00:00:00Z"
          description: "キャンペーン開始日時"
        campaign_end_date:
          type: string
          format: date-time
          example: "2025-03-31T23:59:59Z"
          description: "キャンペーン終了日時"
        pricing:
          type: object
          properties:
            base_fee:
              type: number
              multipleOf: 0.01
              minimum: 0
              example: 2980.00
              description: "基本料金"
            call_fee:
              type: number
              multipleOf: 0.01
              minimum: 0
              example: 0.00
              description: "通話料金（1分あたり）"
            data_fee:
              type: number
              multipleOf: 0.01
              minimum: 0
              example: 550.00
              description: "データ追加料金（1GBあたり）"
            option_fees:
              type: array
              items:
                type: object
                properties:
                  option_id:
                    type: string
                    example: "insurance_001"
                  option_name:
                    type: string
                    example: "端末保証サービス"
                  monthly_fee:
                    type: number
                    multipleOf: 0.01
                    minimum: 0
                    example: 550.00
              description: "オプション料金一覧"
            discounts:
              type: array
              items:
                type: object
                properties:
                  discount_id:
                    type: string
                    example: "family_discount"
                  discount_name:
                    type: string
                    example: "家族割引"
                  discount_type:
                    type: string
                    enum: [percentage, fixed_amount]
                    example: "fixed_amount"
                  discount_value:
                    type: number
                    multipleOf: 0.01
                    example: 500.00
                  conditions:
                    type: object
                    properties:
                      min_family_members:
                        type: integer
                        example: 2
              description: "割引情報一覧"
        created_at:
          type: string
          format: date-time
          example: "2024-12-20T10:30:00Z"
          description: "作成日時"
        updated_at:
          type: string
          format: date-time
          example: "2024-12-20T10:30:00Z"
          description: "更新日時"
...
    Device:
      type: object
      properties:
        id:
          type: string
          example: "device_iphone15_001"
        name:
          type: string
          example: "iPhone 15"
        brand:
          type: string
          example: "Apple"
        category:
          type: string
          enum: [iPhone, Android]
        price_range:
          type: string
          enum: [entry, mid, premium]
        price:
          type: number
          format: decimal
          example: 124800
        colors:
          type: array
          items:
            type: string
        storage_options:
          type: array
          items:
            type: string
        in_stock:
          type: boolean
        image_url:
          type: string
          format: uri
...
    ContractCreateRequest:
      type: object
      required:
        - plan_id
        - device_id
      properties:
        plan_id:
          type: string
        device_id:
          type: string
        device_color:
          type: string
        device_storage:
          type: string
        options:
          type: array
          items:
            type: string
        customer_info:
          $ref: '#/components/schemas/CustomerInfo'
        save_as_draft:
          type: boolean
          default: false
    Contract:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [draft, pending_ekyc, pending_payment, active, cancelled]
        plan:
          $ref: '#/components/schemas/Plan'
        device:
          $ref: '#/components/schemas/Device'
        options:
          type: array
          items:
            $ref: '#/components/schemas/Option'
        customer_info:
          $ref: '#/components/schemas/CustomerInfo'
        total_amount:
          type: number
          format: decimal
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ContractTemplate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: テンプレート名
        description:
          type: string
          description: テンプレート説明
        version:
          type: string
          description: テンプレートバージョン
        format:
          type: string
          enum: [html, pdf]
          description: テンプレート形式
        content:
          type: string
          description: テンプレート内容
        variables:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
                enum: [string, number, date, boolean]
              required:
                type: boolean
              description:
                type: string
        is_active:
          type: boolean
          description: アクティブ状態
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
        updated_by:
          type: string
    ContractTemplateCreateRequest:
      type: object
      required:
        - name
        - format
        - content
      properties:
        name:
          type: string
          description: テンプレート名
        description:
          type: string
          description: テンプレート説明
        format:
          type: string
          enum: [html, pdf]
          description: テンプレート形式
        content:
          type: string
          description: テンプレート内容
        variables:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
                enum: [string, number, date, boolean]
              required:
                type: boolean
              description:
                type: string
    ContractTemplateUpdateRequest:
      type: object
      properties:
        name:
          type: string
          description: テンプレート名
        description:
          type: string
          description: テンプレート説明
        content:
          type: string
          description: テンプレート内容
        variables:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
                enum: [string, number, date, boolean]
              required:
                type: boolean
              description:
                type: string
        change_reason:
          type: string
          description: 変更理由
    ContractGenerationRequest:
      type: object
      required:
        - template_id
      properties:
        template_id:
          type: string
          format: uuid
          description: 使用するテンプレートID
        variables:
          type: object
          description: テンプレート変数の値
          additionalProperties: true
        format:
          type: string
          enum: [html, pdf]
          default: pdf
          description: 生成形式
    ContractGenerationResponse:
      type: object
      properties:
        document_id:
          type: string
          format: uuid
          description: 生成された文書ID
        document_url:
          type: string
          format: uri
          description: 文書アクセスURL
        format:
          type: string
          enum: [html, pdf]
          description: 生成形式
        generated_at:
          type: string
          format: date-time
          description: 生成日時
        expires_at:
          type: string
          format: date-time
          description: URL有効期限
    ElectronicSignatureRequest:
      type: object
      required:
        - signature_method
      properties:
        signature_method:
          type: string
          enum: [embedded, remote, cloud]
          description: 署名方式
        signer_info:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
              format: email
            phone:
              type: string
          required:
            - name
            - email
        return_url:
          type: string
          format: uri
          description: 署名完了後のリダイレクトURL
        webhook_url:
          type: string
          format: uri
          description: 署名完了通知URL
    ElectronicSignatureResponse:
      type: object
      properties:
        signature_id:
          type: string
          format: uuid
          description: 署名ID
        signature_url:
          type: string
          format: uri
          description: 署名実行URL
        status:
          type: string
          enum: [pending, in_progress, completed, failed, expired]
          description: 署名ステータス
        signed_at:
          type: string
          format: date-time
          description: 署名完了日時
        certificate_info:
          type: object
          properties:
            certificate_id:
              type: string
            timestamp:
              type: string
              format: date-time
            tsa_url:
              type: string
              format: uri
            hash_algorithm:
              type: string
        expires_at:
          type: string
          format: date-time
          description: 署名URL有効期限
    ContractDocumentResponse:
      type: object
      properties:
        document_id:
          type: string
          format: uuid
        contract_id:
          type: string
          format: uuid
        document_url:
          type: string
          format: uri
          description: 文書ダウンロードURL
        format:
          type: string
          enum: [pdf, html]
        file_size:
          type: integer
          description: ファイルサイズ（バイト）
        signature_status:
          type: string
          enum: [unsigned, signed, verified]
        created_at:
          type: string
          format: date-time
        signed_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          description: URL有効期限
    ContractAuditTrailResponse:
      type: object
      properties:
        contract_id:
          type: string
          format: uuid
        audit_events:
          type: array
          items:
            type: object
            properties:
              event_id:
                type: string
                format: uuid
              event_type:
                type: string
                enum: [created, updated, generated, signed, viewed, downloaded]
              timestamp:
                type: string
                format: date-time
              user_id:
                type: string
              user_name:
                type: string
              ip_address:
                type: string
              user_agent:
                type: string
              details:
                type: object
                additionalProperties: true
              evidence:
                type: object
                properties:
                  certificate_id:
                    type: string
                  timestamp_token:
                    type: string
                  hash_value:
                    type: string
        generated_at:
          type: string
          format: date-time
    ContractSearchResponse:
      type: object
      properties:
        contracts:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              contract_number:
                type: string
              customer_name:
                type: string
              status:
                type: string
                enum: [draft, pending, signed, active, cancelled]
              created_at:
                type: string
                format: date-time
              signed_at:
                type: string
                format: date-time
              total_amount:
                type: number
                format: decimal
        pagination:
          type: object
          properties:
            current_page:
              type: integer
            total_pages:
              type: integer
            total_count:
              type: integer
            per_page:
              type: integer
    CustomerInfo:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        first_name_kana:
          type: string
        last_name_kana:
          type: string
        birth_date:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        phone:
          type: string
        email:
          type: string
          format: email
        address:
          $ref: '#/components/schemas/Address'
...
  responses:
    BadRequestError:
      description: "リクエストエラー（400）"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error_code: "VALIDATION_ERROR"
            message: "入力内容に不備があります"
            details:
              - field: "email"
                code: "INVALID_FORMAT"
                message: "有効なメールアドレスを入力してください"
            request_id: "req_12345678-1234-1234-1234-123456789012"
            timestamp: "2024-12-20T10:30:00Z"
    UnauthorizedError:
      description: "認証エラー（401）"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error_code: "UNAUTHORIZED"
            message: "認証が必要です"
            request_id: "req_12345678-1234-1234-1234-123456789012"
            timestamp: "2024-12-20T10:30:00Z"
...
    NotFoundError:
      description: "リソースが見つかりません（404）"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error_code: "NOT_FOUND"
            message: "指定されたリソースが見つかりません"
...